Viewport paths are now used instead of viewport names.
We still perform name mangling, as a result, a few functions have been
created in JS for this purpose. When writing out to a JS file, we also
include the <script> tag for the utils JS file which was erroneously
ommitted. Coordinates files and clipping paths now only refer to the
*canonical viewport name*, thus reducing redundancy.


git-svn-id: svn://svn.r-forge.r-project.org/svnroot/gridsvg@125 edb9625f-4e0d-4859-8d74-9fd3b1da38cb